+++ app/src/test/java/org/simple/clinic/activity/TheActivityControllerTest.kt
@@ -1,6 +1,7 @@
 package org.simple.clinic.activity
 
 import com.google.common.truth.Truth.assertThat
+import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.clearInvocations
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.never
@@ -16,6 +17,9 @@ import org.junit.After
 import org.junit.Rule
 import org.junit.Test
 import org.simple.clinic.TestData
+import org.simple.clinic.forgotpin.createnewpin.ForgotPinCreateNewPinScreenKey
+import org.simple.clinic.home.HomeScreenKey
+import org.simple.clinic.login.applock.AppLockScreenKey
 import org.simple.clinic.main.TheActivityEffect
 import org.simple.clinic.main.TheActivityEffectHandler
 import org.simple.clinic.main.TheActivityEvent
@@ -24,6 +28,7 @@ import org.simple.clinic.main.TheActivityModel
 import org.simple.clinic.main.TheActivityUi
 import org.simple.clinic.main.TheActivityUiRenderer
 import org.simple.clinic.main.TheActivityUpdate
+import org.simple.clinic.navigation.v2.compat.wrap
 import org.simple.clinic.patient.PatientRepository
 import org.simple.clinic.storage.MemoryValue
 import org.simple.clinic.user.User
@@ -33,7 +38,6 @@ import org.simple.clinic.user.User.LoggedInStatus.RESETTING_PIN
 import org.simple.clinic.user.User.LoggedInStatus.RESET_PIN_REQUESTED
 import org.simple.clinic.user.UserSession
 import org.simple.clinic.user.UserStatus
-import java.util.Optional
 import org.simple.clinic.util.RxErrorsRule
 import org.simple.clinic.util.TestUtcClock
 import org.simple.clinic.util.scheduler.TestSchedulersProvider
@@ -42,6 +46,7 @@ import org.simple.clinic.widgets.UiEvent
 import org.simple.mobius.migration.MobiusTestFixture
 import java.time.Duration
 import java.time.Instant
+import java.util.Optional
 import java.util.UUID
 import java.util.concurrent.TimeUnit
 
@@ -59,6 +64,11 @@ class TheActivityControllerTest {
 
   private val currentTimestamp = Instant.parse("2018-01-01T00:00:00Z")
   private val clock = TestUtcClock(currentTimestamp)
+  private val user = TestData.loggedInUser(
+      uuid = UUID.fromString("049ee3e0-f5a8-4ba6-9270-b20231d3fe50"),
+      loggedInStatus = LOGGED_IN,
+      status = UserStatus.ApprovedForSyncing
+  )
 
   private lateinit var testFixture: MobiusTestFixture<TheActivityModel, TheActivityEvent, TheActivityEffect>
 
@@ -68,41 +78,24 @@ class TheActivityControllerTest {
   }
 
   @Test
-  fun `when activity is started and user is logged out then app lock shouldn't be shown`() {
-    // when
-    setupController()
-
-    // then
-    verify(ui, never()).showAppLockScreen()
-    verifyNoMoreInteractions(ui)
-  }
-
-  @Test
   fun `when activity is started, user has requested an OTP, and user was inactive then app lock should be shown`() {
     // given
-    whenever(userSession.loggedInUserImmediate()).thenReturn(TestData.loggedInUser(
-        uuid = UUID.fromString("049ee3e0-f5a8-4ba6-9270-b20231d3fe50"),
-        loggedInStatus = OTP_REQUESTED,
-        status = UserStatus.ApprovedForSyncing
-    ))
+
+    whenever(userSession.loggedInUserImmediate()).thenReturn(user.otpRequested())
     val lockAfterTime = currentTimestamp.minusSeconds(TimeUnit.MINUTES.toSeconds(1))
 
     // when
     setupController(lockAtTime = lockAfterTime)
 
     // then
-    verify(ui).showAppLockScreen()
+    verify(ui).showInitialScreen(AppLockScreenKey(HomeScreenKey))
     verifyNoMoreInteractions(ui)
   }
 
   @Test
   fun `when activity is started, user is logged in, and user was inactive then app lock should be shown`() {
     // given
-    whenever(userSession.loggedInUserImmediate()).thenReturn(TestData.loggedInUser(
-        uuid = UUID.fromString("049ee3e0-f5a8-4ba6-9270-b20231d3fe50"),
-        loggedInStatus = LOGGED_IN,
-        status = UserStatus.ApprovedForSyncing
-    ))
+    whenever(userSession.loggedInUserImmediate()).thenReturn(user)
 
     val lockAfterTime = currentTimestamp.minusSeconds(TimeUnit.MINUTES.toSeconds(1))
 
@@ -110,36 +103,28 @@ class TheActivityControllerTest {
     setupController(lockAtTime = lockAfterTime)
 
     // then
-    verify(ui).showAppLockScreen()
+    verify(ui).showInitialScreen(AppLockScreenKey(HomeScreenKey))
     verifyNoMoreInteractions(ui)
   }
 
   @Test
   fun `when activity is started, user has requested a PIN reset, and user was inactive then app lock should be shown`() {
     // given
-    whenever(userSession.loggedInUserImmediate()).thenReturn(TestData.loggedInUser(
-        uuid = UUID.fromString("049ee3e0-f5a8-4ba6-9270-b20231d3fe50"),
-        loggedInStatus = RESET_PIN_REQUESTED,
-        status = UserStatus.ApprovedForSyncing
-    ))
+    whenever(userSession.loggedInUserImmediate()).thenReturn(user.resetPinRequested())
     val lockAfterTime = currentTimestamp.minusSeconds(TimeUnit.MINUTES.toSeconds(1))
 
     // when
     setupController(lockAtTime = lockAfterTime)
 
     // then
-    verify(ui).showAppLockScreen()
+    verify(ui).showInitialScreen(AppLockScreenKey(HomeScreenKey))
     verifyNoMoreInteractions(ui)
   }
 
   @Test
   fun `when activity is started, user is resetting the PIN, and user was inactive then app lock should not be shown`() {
     // given
-    whenever(userSession.loggedInUserImmediate()).thenReturn(TestData.loggedInUser(
-        uuid = UUID.fromString("049ee3e0-f5a8-4ba6-9270-b20231d3fe50"),
-        loggedInStatus = RESETTING_PIN,
-        status = UserStatus.ApprovedForSyncing
-    ))
+    whenever(userSession.loggedInUserImmediate()).thenReturn(user.resettingPin())
 
     val lockAfterTime = currentTimestamp.minusSeconds(TimeUnit.MINUTES.toSeconds(1))
 
@@ -147,18 +132,14 @@ class TheActivityControllerTest {
     setupController(lockAtTime = lockAfterTime)
 
     // then
-    verify(ui, never()).showAppLockScreen()
+    verify(ui).showInitialScreen(ForgotPinCreateNewPinScreenKey().wrap())
     verifyNoMoreInteractions(ui)
   }
 
   @Test
   fun `when app is started unlocked and lock timer hasn't expired yet then the timer should be unset`() {
     // given
-    whenever(userSession.loggedInUserImmediate()).thenReturn(TestData.loggedInUser(
-        uuid = UUID.fromString("049ee3e0-f5a8-4ba6-9270-b20231d3fe50"),
-        loggedInStatus = LOGGED_IN,
-        status = UserStatus.ApprovedForSyncing
-    ))
+    whenever(userSession.loggedInUserImmediate()).thenReturn(user)
 
     val lockAfterTimestamp = MemoryValue(
         defaultValue = Optional.empty(),
@@ -170,17 +151,14 @@ class TheActivityControllerTest {
 
     // then
     assertThat(lockAfterTimestamp.hasValue).isFalse()
+    verify(ui).showInitialScreen(HomeScreenKey)
     verifyNoMoreInteractions(ui)
   }
 
   @Test
   fun `when app is started locked and lock timer hasn't expired yet then the timer should not be unset`() {
     // given
-    whenever(userSession.loggedInUserImmediate()).thenReturn(TestData.loggedInUser(
-        uuid = UUID.fromString("049ee3e0-f5a8-4ba6-9270-b20231d3fe50"),
-        loggedInStatus = LOGGED_IN,
-        status = UserStatus.ApprovedForSyncing
-    ))
+    whenever(userSession.loggedInUserImmediate()).thenReturn(user)
     val lockAfterTimestamp = MemoryValue(
         defaultValue = Optional.empty(),
         currentValue = Optional.of(currentTimestamp.minusSeconds(TimeUnit.MINUTES.toSeconds(5)))
@@ -191,23 +169,18 @@ class TheActivityControllerTest {
 
     // then
     assertThat(lockAfterTimestamp.hasValue).isTrue()
-    verify(ui).showAppLockScreen()
+    verify(ui).showInitialScreen(AppLockScreenKey(HomeScreenKey))
     verifyNoMoreInteractions(ui)
   }
 
   @Test
   fun `the logged out alert must be shown only at the instant when a user gets verified for login`() {
     // given
-    val user = TestData.loggedInUser(
-        uuid = UUID.fromString("bed4a670-7f03-44ab-87ca-f297ca35375a"),
-        status = UserStatus.ApprovedForSyncing,
-        loggedInStatus = OTP_REQUESTED
-    )
-
+    whenever(userSession.loggedInUserImmediate()).thenReturn(user)
     val userStream: Observable<Optional<User>> = Observable.just(
+        Optional.of(user.otpRequested()),
         Optional.of(user),
-        Optional.of(user.copy(loggedInStatus = LOGGED_IN)),
-        Optional.of(user.copy(loggedInStatus = LOGGED_IN))
+        Optional.of(user)
     )
 
     // when
@@ -217,6 +190,7 @@ class TheActivityControllerTest {
     )
 
     // then
+    verify(ui).showInitialScreen(HomeScreenKey)
     verify(ui).showUserLoggedOutOnOtherDeviceAlert()
     verifyNoMoreInteractions(ui)
   }
@@ -224,22 +198,19 @@ class TheActivityControllerTest {
   @Test
   fun `the logged out alert must not be shown if the user is already logged in when the screen is opened`() {
     // given
-    val user = TestData.loggedInUser(
-        uuid = UUID.fromString("bed4a670-7f03-44ab-87ca-f297ca35375a"),
-        status = UserStatus.ApprovedForSyncing,
-        loggedInStatus = LOGGED_IN
-    )
+    whenever(userSession.loggedInUserImmediate()).thenReturn(user)
     whenever(userSession.loggedInUser()).thenReturn(
         Observable.just(
             Optional.of(user),
-            Optional.of(user.copy(loggedInStatus = LOGGED_IN)),
-            Optional.of(user.copy(loggedInStatus = LOGGED_IN)))
+            Optional.of(user),
+            Optional.of(user))
     )
 
     // when
     setupController()
 
     // then
+    verify(ui).showInitialScreen(AppLockScreenKey(HomeScreenKey))
     verify(ui, never()).showUserLoggedOutOnOtherDeviceAlert()
     verifyNoMoreInteractions(ui)
   }
@@ -248,15 +219,10 @@ class TheActivityControllerTest {
   fun `when user is denied access then access denied screen should show`() {
     //given
     val fullName = "Anish Acharya"
-    val loggedInUser = TestData.loggedInUser(
-        uuid = UUID.fromString("0b350f89-ed0e-4922-b384-7f7a9bf3aba0"),
-        name = fullName,
-        status = UserStatus.DisapprovedForSyncing,
-        loggedInStatus = LOGGED_IN
-    )
-    whenever(userSession.loggedInUser()).thenReturn(Observable.just(loggedInUser.toOptional()))
-    whenever(patientRepository.clearPatientData()).thenReturn(Completable.complete())
+    val loggedInUser = user
+        .withFullName(fullName)
     whenever(userSession.loggedInUserImmediate()).thenReturn(loggedInUser)
+    whenever(patientRepository.clearPatientData()).thenReturn(Completable.complete())
     val userDisapprovedSubject = PublishSubject.create<Boolean>()
 
     //when
@@ -264,9 +230,15 @@ class TheActivityControllerTest {
         userDisapprovedStream = userDisapprovedSubject,
         lockAtTime = Instant.now(clock)
     )
+    // then
+    verify(ui).showInitialScreen(HomeScreenKey)
+    verifyNoMoreInteractions(ui)
+    verify(patientRepository, never()).clearPatientData()
+
+    // when
     userDisapprovedSubject.onNext(true)
 
-    //then
+    // then
     verify(patientRepository).clearPatientData()
     verify(ui).showAccessDeniedScreen(fullName)
     verifyNoMoreInteractions(ui)
@@ -275,20 +247,14 @@ class TheActivityControllerTest {
   @Test
   fun `when user has access then the access denied screen should not appear`() {
     //given
-    val fullName = "Anish Acharya"
-    val loggedInUser = TestData.loggedInUser(
-        uuid = UUID.fromString("0b350f89-ed0e-4922-b384-7f7a9bf3aba0"),
-        name = fullName,
-        status = UserStatus.ApprovedForSyncing,
-        loggedInStatus = LOGGED_IN
-    )
-    whenever(userSession.loggedInUser()).thenReturn(Observable.just(loggedInUser.toOptional()))
+    whenever(userSession.loggedInUserImmediate()).thenReturn(user)
+    whenever(userSession.loggedInUser()).thenReturn(Observable.just(user.toOptional()))
 
     //when
     setupController(lockAtTime = Instant.now(clock))
 
     //then
-    verify(ui, never()).showAccessDeniedScreen(fullName)
+    verify(ui).showInitialScreen(HomeScreenKey)
     verifyNoMoreInteractions(ui)
     verify(patientRepository, never()).clearPatientData()
   }
@@ -296,6 +262,7 @@ class TheActivityControllerTest {
   @Test
   fun `the sign in screen must be shown only at the moment where the user gets logged out`() {
     // given
+    whenever(userSession.loggedInUserImmediate()).thenReturn(user)
     val userUnauthorizedSubject = PublishSubject.create<Boolean>()
 
     // when
@@ -349,7 +316,7 @@ class TheActivityControllerTest {
   }
 
   private fun setupController(
-      userStream: Observable<Optional<User>> = Observable.just(Optional.empty()),
+      userStream: Observable<Optional<User>> = Observable.just(Optional.of(user)),
       userUnauthorizedStream: Observable<Boolean> = Observable.just(false),
       userDisapprovedStream: Observable<Boolean> = Observable.just(false),
       lockAfterTimestamp: MemoryValue<Optional<Instant>>
@@ -378,4 +345,12 @@ class TheActivityControllerTest {
     )
     testFixture.start()
   }
+
+  private fun User.otpRequested(): User = copy(loggedInStatus = OTP_REQUESTED)
+
+  private fun User.resetPinRequested(): User = copy(loggedInStatus = RESET_PIN_REQUESTED)
+
+  private fun User.resettingPin(): User = copy(loggedInStatus = RESETTING_PIN)
+
+  private fun User.disapprovedForSyncing(): User = copy(status = UserStatus.DisapprovedForSyncing)
 }