+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `.gitattributes` for specifying eol for `strings.xml`  
 - Support `ContactPatientSheet` with no appointment
 - Migrate to Gradle Version Catalog  
+- Move logic for deciding the initial screen to the Mobius loop in `TheActivity`
 - [In Progress: 04 Aug 2021] Add support for Sri Lanka 
 
 ### Changes